Design Curves Files (.dcv) for 4 WSUD Systems
Data collated from MUSIC models to create design curves for sizing four
types of WSUD systems. The files are comma-separated-variable, but have
been given the .dcv extension that UrbanBEATS will work with when
dealing with design curves.
Note: .dcv specification may change in future, but this is the first
pass.