Begin of Retrofit Algorithm and Bug Fix
First pieces of code written for the retrofit algorithm, GUI revised for
this purpose, new elements added.
Minor bug fix in techstrategy_eval that was causing lookup of an array
to crash due to an out-of-range index.
Updated __init__.py to include the new techretrofit.py module that will
contain all the retrofit code.